+++ test/Sentry.Tests/SentryOptionsTests.Integrations_default_ones_are_properly_registered.Core.verified.txt
