+++ test/Sentry.Tests/SentryOptionsTests.Integrations_default_ones_are_properly_registered.DotNet.verified.txt
